+++ App.js
@@ -4,14 +4,14 @@ import { StackNavigator } from 'react-navigation'
 import LinearGradient from 'react-native-linear-gradient'
 
 //Screens
-import HomeScreen from './screens/HomeScreen'
+import MainScreen from './screens/MainScreen'
 import UploadScreen from './screens/UploadScreen'
 import SplashScreen from './screens/SplashScreen'
 
 const RootNavigator = StackNavigator({
   Splash: { screen: SplashScreen },
   Upload: { screen: UploadScreen }, // Home should be on this line!!!
-  Home: { screen: HomeScreen }
+  Main: { screen: MainScreen }
   },
   { headerMode: 'none' },
   {initialRoute: 'Splash'}
@@ -23,7 +23,7 @@ export default class App extends Component {
       <View style={{flex: 1}}>
       <StatusBar 
             barStyle = "light-content"
-            backgroundColor="#0084FF"
+            backgroundColor="black"
             hidden = {false}/>
       <RootNavigator />
      </View>